+++ src/lib/components/monoEmoji.ts
@@ -1,0 +1,3 @@
+export default function monoEmoij(emoji: string) {
+	return emoji.replaceAll('\uFE0F', '') + '\uFE0E';
+}