+++ src/lib/data/getTransactionVenues.ts
@@ -1,13 +1,14 @@
+import type { Database } from '$data/database';
 import type { TransactionRow } from '$data/types';
 import type { SupabaseClient } from '@supabase/supabase-js';
 
-export default async function getVenueTitles(
-	supabase: SupabaseClient,
+export default async function getTransactionVenues(
+	supabase: SupabaseClient<Database>,
 	transactions: TransactionRow[]
 ) {
 	return await supabase
 		.from('venues')
-		.select('id, title')
+		.select('*')
 		.in(
 			'id',
 			Array.from(