+++ src/routes/scholar/[id]/transactions/+page.ts
@@ -1,4 +1,5 @@
-import getVenueTitles from '$lib/data/venueTitles';
+import getTransactionCurrencies from '$lib/data/getTransactionCurrencies';
+import getTransactionVenues from '$lib/data/getTransactionVenues';
 import type { PageLoad } from './$types';
 
 export const load: PageLoad = async ({ parent, params }) => {
@@ -15,11 +16,18 @@ export const load: PageLoad = async ({ parent, params }) => {
 	const { data: venues, error: venueError } =
 		transactions === null
 			? { data: null, error: null }
-			: await getVenueTitles(supabase, transactions);
+			: await getTransactionVenues(supabase, transactions);
 	if (venueError) console.log(venueError);
+
+	const { data: currencies, error: currencyError } =
+		transactions === null
+			? { data: null, error: null }
+			: await getTransactionCurrencies(supabase, transactions);
+	if (currencyError) console.log(currencyError);
 
 	return {
 		transactions,
-		venues
+		venues,
+		currencies
 	};
 };