+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Dates should be in`YYYY-MM-DD` format and versions are in [semantic versioning](
 - Show transactions status on submission page.
 - Added a toggleable submission status to each submission, to mark when it is no longer in review.
 - Show submissions on scholar page.
+- Transaction cancelation by giver, editor, or minter.
 
 ### Fixed